refactor: split over-cap orchestrator core + task-index store under the file-size cap
Both files only passed CI via file-size exemptions. Behaviour-preserving split with an
unchanged public interface: store.ts's pure lock-retry + on-disk codec move to
index/store-codec.ts; core.ts's git-worktree side-effects move to a WorktreeCoordinator
collaborator, its in-place task-field edits to a TaskEditor collaborator, and its pure
path/repo-key helpers to core-helpers.ts. The Orchestrator and TaskIndexStore classes keep
every public method as a thin delegator, so no caller changed.
